transactions: separate testing routines in a new submodule
So we can reuse them from the fuzz targets.
